Changed ints to floats to include truncation from device side (#158)
* Update injector.py

* Update injector.py

* Update mysql_tools.py

* Update injector.py

* Update mysql_tools.py

* Added CO2 sensor stuff into classify_request

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update mysql_tools.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update injector.py

* Update mysql_tools.py

* Update injector.py

* Update injector.py

* Update mysql_tools.py

* Update mysql_tools.py

* Update mysql_tools.py

* Added the main info for the co2 sensor into the injector

* Added the tools for the co2 sensor

* Added missing info to the injector file

* Attemping to change a weird error

* Fixed that weird error this time

* Changed data type for the test co2 data

* Just making sure of something

* another quick test

* Trying to make the length testing more robust

* Trying to make the length testing more robust

* Trying to make the length testing more robust

* Made the AQ and CO2 request classifier work hopefully in more realistic conditions

* Added the code for the Weather sensor

* Added a quick help section on the argument parser

* Forgot the num_weather_fields argument

* Put a comma instead of a decimal point

* Changed variable names from 'co2_conc' to 'co2' and from 'uv_index' to 'noise'

* Changed 'co2' to 'co2_ppm' because I misread the comment

* Changed 'co2' to 'co2_ppm' because I misread the comment

* Changed the name of the database to match the actual sql database

* Removed 'noise' from beginning

* Changed 'error_flag' to 'errorFlag'

* Wrong error flag in global

* Added test case

* Changed data type size

* Changed data type to int for now

* Removed print statement

* Changed data types and type indictaion numbers

* Changed test data a little bit

* Changed quick data again

* The problem wasn't data it was the recognition of type of data

+++ mysql/mysql_tools.py
@@ -261,14 +261,14 @@ class SQLObject:
                     'Spectrum length is {}, should be 4096'.format(
                         len(data['spectrum'])))
         elif packettype == 'AQ':
-            data_types = {'hash': str, 'stationID': int, 'oneMicron': int, 'twoPointFiveMicron':
-                          int, 'tenMicron': int, 'error_flag': int}
+            data_types = {'hash': str, 'stationID': int, 'oneMicron': float, 'twoPointFiveMicron':
+                          float, 'tenMicron': float, 'error_flag': int}
         elif packettype == 'CO2':
-            data_types = {'hash': str, 'stationID': int, 'co2_ppm': int, 'noise':
-                          int, 'error_flag': int}
+            data_types = {'hash': str, 'stationID': int, 'co2_ppm': float, 'noise':
+                          float, 'error_flag': int}
         elif packettype == 'Weather':
-            data_types = {'hash': str, 'stationID': int, 'temperature': int, 'pressure':
-                          int, 'humidity': int, 'error_flag': int}
+            data_types = {'hash': str, 'stationID': int, 'temperature': float, 'pressure':
+                          float, 'humidity': float, 'error_flag': int}
         elif packettype == 'log':
             data_types = {'hash': str, 'stationID': int, 'msgCode': int,
                           'msgText': str}
